Refactor types.tsx
