+++ sematic/ui/src/types/none.tsx
@@ -1,0 +1,5 @@
+import { Chip } from '@mui/material';
+
+export default function NoneValueView() {
+    return <Chip label="NONE" size="small" />;
+}
